+++ Changelog.md
@@ -1,5 +1,7 @@
 # Version UPCOMING (...)
 
+* Adds a `cache` Cargo feature that provides `cache::StatementCache` for caching prepared
+  statements.
 * Adds `insert` convenience method to `Statement` which returns the row ID of an inserted row.
 * Adds `exists` convenience method returning whether a query finds one or more rows.
 * Adds support for serializing types from the `serde_json` crate. Requires the `serde_json` feature.